Classe ProfessorValidator

+++ src/main/java/sgallalucas/fitlab/services/ProfessorService.java
@@ -1,10 +1,14 @@
 package sgallalucas.fitlab.services;
 
+import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import sgallalucas.fitlab.dtos.ProfessorDto;
+import sgallalucas.fitlab.exceptions.NotAllowedOperationException;
 import sgallalucas.fitlab.model.Professor;
 import sgallalucas.fitlab.repositories.ProfessorRepository;
+import sgallalucas.fitlab.repositories.WorkoutRepository;
+import sgallalucas.fitlab.validators.ProfessorValidator;
 
 import java.util.List;
 import java.util.UUID;
@@ -14,14 +18,17 @@ import java.util.UUID;
 public class ProfessorService {
 
     private final ProfessorRepository repository;
+    private final ProfessorValidator validator;
+    private final WorkoutRepository workoutRepository;
 
     public Professor save(Professor professor) {
+        validator.validation(professor.getEmail());
         return repository.save(professor);
     }
 
     public Professor findById(UUID uuid) {
         return repository.findById(uuid)
-                .orElseThrow(() -> new RuntimeException("Professor not found"));
+                .orElseThrow(() -> new EntityNotFoundException("Professor not found"));
     }
 
     public List<Professor> findAll() {
@@ -29,11 +36,19 @@ public class ProfessorService {
     }
 
     public void update(Professor professor) {
+        validator.validation(professor.getEmail());
         repository.save(professor);
     }
 
     public void delete(Professor professor) {
+        if (hasWorkout(professor)) {
+            throw new NotAllowedOperationException("It is not allowed delete a professor who has workouts");
+        }
         repository.delete(professor);
+    }
+
+    public boolean hasWorkout(Professor professor) {
+        return workoutRepository.existsByProfessor(professor);
     }
 
     public Professor convertToEntity(ProfessorDto dto) {
